sync function

+++ index.js
@@ -10,16 +10,24 @@ const app = express();
 app.use(cors());
 app.use(express.json());
 
-connection.authenticate().then(() => {
-  console.log(`Connected to database - ${process.env.DB_NAME}`);
+connection
+  .authenticate()
+  .then(() => {
+    if (process.env.SYNC) {
+      connection.sync({ force: true });
+      process.exit();
+    } else {
+      console.log(`Connected to database - ${process.env.DB_NAME}`);
 
-  app.use("/auth", require("./routes/auth"));
-  app.use("/orders", require("./routes/orders"));
-  app.use("/products", require("./routes/products"));
+      app.use("/auth", require("./routes/auth"));
+      app.use("/orders", require("./routes/orders"));
+      app.use("/products", require("./routes/products"));
 
-  connection.sync({ force: true });
-
-  app.listen(process.env.PORT, () => {
-    console.log(`App listening @${process.env.PORT}`);
+      app.listen(process.env.PORT, () => {
+        console.log(`App listening @${process.env.PORT}`);
+      });
+    }
+  })
+  .catch((err) => {
+    console.error(err);
   });
-});
